move file contents to right file fix a display error
